# Handling for GetAsyncRunResults failures in the callback